requirements.txt
flask==2.3.2
xgboost==1.7.6
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.0.3
kafka-python==2.0.2
matplotlib==3.7.2
shap==0.41.0